In the last series of commits, I converted the previous HTML code used to create the legend into TypeScript.
To do so, I created a function called genSVGArrow which takes in several arrow attributes (color, dashed, arrow name, etc.) as parameters and dynamically creates the SVG elements needed to generate the arrows (line, marker, deft, etc.). I also created multiple helper functions to dynamically generate the HTML table:

getTableElement takes in the arrow name as input and returns the arrow with its proper attributes
getTableText takes in the arrow name as input and returns the HTML table element td which the inner text equal to the input
To continue, I created several helper functions to append each of the elements to the table:

appendArrow takes in the arrow name as input and returns the HTML table element td with the arrow appended onto it
appendTableElement takes in an element name and dynamically creates the SVG elements that are not arrows (i.e. boxes representing modules/sensors)
appendIMG dynamically creates the img tag and appends it to the td element of the table
Finally, the initialize function intializes the table elements (tbody, trTop, trBottom, etc.), appends the arrow, table text, and other SVG elements, and returns the complete table.